Add files via upload
This project proposes the Ridge Regression model for the prediction of the price of used cars in the Indian market. The model uses a Ridge Regression algorithm and is trained on various attributes of used cars such as brand, model, year, fuel type, transmission, ownership, engine capacity, power, seats, and more. The model has high accuracy and can predict the prices of used cars very close to their original values.